refactored svg asset names
