+++ src/javascript/app_2/App/Assets/Common/icon_arrow.jsx
