+++ src/javascript/app_2/App/Assets/Common/index.js
@@ -1,2 +1,3 @@
-export * from './arrowback.jsx';
-export * from './arrowhead.jsx';
+export * from './icon_back.jsx';
+export * from './icon_close.jsx';
+export * from './icon_arrow.jsx';
